+++ update_finish.sh
@@ -40,9 +40,6 @@ timedatectl set-ntp true
 # fix factory reset issue
 cp /home/pi/sisbot-server/sisbot/factory_reset.sh /home/pi/sisbot-server/
 
-# update rc.local
-cp /home/pi/sisbot-server/sisbot/rc_fix.local /etc/rc.local
-
 # make sure log file location existsSync
 mkdir -p /var/log/sisyphus